Add basic command understanding wrapper package
Imports the command understanding python module and makes it available
to catkin packages. It may be difficult to use because it's Python 3 only
though.
Included a basic ROS + parser proof of concept node running in Python 3
Ideally, other code would just import the module, but it may be necessary
to talk to the parser over ROS to avoid having to run the importing code
in Python 3