Change submodule include to git pip requirement
The pip requirement route needs the dependency to have a setup.py, but it lets us move the dependency resolution straight into catkin_virtualenv and the devel space.
Beacause the resulting venv is per-package, the module isn't exposed to the rest of the workspace.
In this Python2/3 situation, we don't want the rest of the workspace to import the module, so this is fine.
I'm not sure about these tradeoff, but we'll try this for now